pyproject.toml
[project]
name = "terminals"
version = "0.1.0"
description = "Multi-tenant terminal orchestrator for Open Terminal."
readme = "README.md"
authors = [
    { name = "Timothy Jaeryang Baek", email = "tim@openwebui.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "click>=8.1.0",
    "httpx>=0.27.0",
    "aiodocker>=0.23.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.15.0",
    "pydantic-settings>=2.0.0",
    "websockets>=13.0",
    "loguru>=0.7.3",
    "kubernetes-asyncio>=31.0.0",
    "cryptography>=44.0.0",
]

[project.scripts]
terminals = "terminals.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["terminals"]
